+++ docs/source/extending/test_myjitmodule2.py
